Reverted last commit changes

+++ models/results.go
@@ -91,7 +91,7 @@ func (set ResultSet) Avg() (float64, error) {
 	return sum / float64(len(set.Results)), nil
 }
 
-func (set ResultSet) NumericLast(Count float64) (ResultSet, error) {
+func (set ResultSet) NumericLast(Count int) (ResultSet, error) {
 	var dummyResultSet ResultSet
 	dummyResultSet.Results = make([]Result, 0)
 
@@ -110,7 +110,7 @@ func (set ResultSet) NumericLast(Count float64) (ResultSet, error) {
 	}
 
 	indexPointer := len(set.Results)
-	indexTarget := indexPointer - int(Count)
+	indexTarget := indexPointer - Count
 
 	//If we request more items than present in the result set, just return everything we have
 	if indexTarget < 0 {
